pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "soc-analyst-env"
version = "1.0.0"
description = "SOC Analyst OpenEnv Environment for cybersecurity alert triage, investigation, and correlation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Abe Boodala"},
]

dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["soc_analyst_env*", "server*", "models*", "client*"]
